Changes in indexes. Modified 'fields_to_change' and created new indexes

+++ airflow/data_analysis/classify_elastic/structured_data/index_infraestructuras_generales/index_guaguas_viajeros.py
@@ -23,7 +23,7 @@ table_start_and_end = {
     "start_value_row": 22,
     "start_value_col": 1
 }
-type_value = int
+type_value = float
 
 attribute_to_split = [{
     "attributes":["year","month"],      #Attributes to split
